Split rust emitter into several files
